chore(admissions): move applicant toggle to own file
