+++ src/modules/admissions/CloseAdmission/CloseAdmission.tsx
@@ -1,12 +1,10 @@
-import { useMutation, useQuery } from '@apollo/client'
+import { useQuery } from '@apollo/client'
 import { FullPageError } from 'components/FullPageComponents'
 import { FullContentLoader } from 'components/Loading'
-import { useState } from 'react'
-import { toast } from 'react-hot-toast'
 import styled from 'styled-components'
+import { VALID_APPLICANTS_QUERY } from '../AdmissionDashboard/queries'
 import { ApplicantNode } from '../types'
-import { TOGGLE_APPLICANT_WILL_BE_ADMITTED_MUTATION } from './mutations'
-import { VALID_APPLICANTS_QUERY } from './queries'
+import { ToggleApplicantInline } from './ToggleApplicantInline'
 const Wrapper = styled.div`
   display: flex;
   flex-direction: column;
@@ -36,68 +34,6 @@ const TableRow = styled.div`
 const TableCell = styled.div`
   width: 170px;
 `
-
-interface ToggleApplicantWillBeAdmittedReturns {
-  toggleApplicantWillBeAdmitted: {
-    success: boolean
-  }
-}
-
-interface CloseAdmissionApplicantInlineProps {
-  applicant: ApplicantNode
-}
-
-const CloseAdmissionApplicantInline: React.VFC<
-  CloseAdmissionApplicantInlineProps
-> = ({ applicant }) => {
-  const [willBeAdmitted, setWillBeAdmitted] = useState(applicant.willBeAdmitted)
-
-  const [toggleApplicant] = useMutation<ToggleApplicantWillBeAdmittedReturns>(
-    TOGGLE_APPLICANT_WILL_BE_ADMITTED_MUTATION,
-    {
-      variables: { id: applicant.id },
-      onCompleted({ toggleApplicantWillBeAdmitted }) {
-        const { success } = toggleApplicantWillBeAdmitted
-
-        if (success) {
-          setWillBeAdmitted(!willBeAdmitted)
-          return
-        }
-        toast.error('No gikk galt')
-      },
-    }
-  )
-
-  const handleToggleApplicant = () => {
-    const toggleFallback = willBeAdmitted
-    setWillBeAdmitted(!toggleApplicant)
-    toggleApplicant().then(({ data }) => {
-      if (data === null || data === undefined) {
-        setWillBeAdmitted(toggleFallback)
-        return
-      }
-      const {
-        toggleApplicantWillBeAdmitted: { success },
-      } = data
-
-      if (!success) {
-        setWillBeAdmitted(toggleFallback)
-        return
-      }
-    })
-  }
-
-  return (
-    <div key={applicant.id}>
-      <span>{applicant.fullName} </span>
-      <input
-        type="checkbox"
-        checked={willBeAdmitted}
-        onChange={handleToggleApplicant}
-      />
-    </div>
-  )
-}
 
 export const CloseAdmission: React.VFC = () => {
   const { error, loading, data } = useQuery(VALID_APPLICANTS_QUERY)
@@ -144,9 +80,8 @@ export const CloseAdmission: React.VFC = () => {
               {applicant.priorities[2]?.internalGroupPriority}
             </TableCell>
             <TableCell>
-              <input type="checkbox"></input>
+              <ToggleApplicantInline applicant={applicant} />
             </TableCell>
-            {/* <CloseAdmissionApplicantInline applicant={applicant} /> */}
           </TableRow>
         ))}
       </Table>
